implement crate save capability
commit fd2a9a457d86b2f7d6327d2392dd8e79f36b011e
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Dec 11 14:27:18 2020 +1100

    implement methods to update crate data on change and save it back

commit 706c1818f94b061cc25500bb8e0da9cb56cf630f
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Dec 11 14:21:36 2020 +1100

    add collectionId to suit revised api

commit 975c81936886b08d6e568af62cd77f39fac72421
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Dec 11 14:20:47 2020 +1100

    revise return from remove prop and entity; formatting

commit 17e4b70fa91b535b6724498c261f3500d3258dbd
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Dec 11 14:19:46 2020 +1100

    add property x-testing to headers and don't save crate if true

    Saving a crate requires a onedrive backend that isn't available in
    testing.

commit 71d57c102da2cdf8bc97208b0e668475346b4e1f
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Dec 11 14:02:28 2020 +1100

    wire up update and save crate capability

+++ api/src/lib/entities.js
@@ -43,17 +43,18 @@ export async function insertEntity({ entity, collectionId }) {
     }
 }
 
-export async function updateEntity({ entityId, name, eid }) {
-    let entity = await models.entity.findOne({ where: { id: entityId } });
-
+export async function updateEntity({ collectionId, entityId, name, eid }) {
+    let entity = await getEntity({ id: entityId, collectionId });
     if (!entity) {
-        throw new Error(`Unable to find an entity with that id`);
+        throw new Error(`You don't have permission to access that entity`);
     }
 
     let update = {};
     if (name) update.name = name;
     if (eid) update.eid = eid;
-    return (await entity.update(update)).get();
+
+    entity = (await entity.update(update)).get();
+    return entity;
 }
 
 export async function attachProperty({ collectionId, entityId, property, value, typeDefinition }) {
@@ -62,12 +63,13 @@ export async function attachProperty({ collectionId, entityId, property, value, 
         throw new Error(`You don't have permission to access that entity`);
     }
     let fqname = typeDefinition?.id ? typeDefinition.id : "";
-    return await models.property.create({
+    property = await models.property.create({
         name: property,
         definition: { id: fqname },
         value,
         entityId,
     });
+    return property;
 }
 
 export async function updateProperty({ collectionId, entityId, propertyId, value }) {
@@ -79,7 +81,8 @@ export async function updateProperty({ collectionId, entityId, propertyId, value
         where: { id: propertyId },
     });
     property.value = value;
-    return await property.save();
+    property = await property.save();
+    return property;
 }
 
 export async function removeProperty({ collectionId, entityId, propertyId }) {
@@ -96,6 +99,7 @@ export async function removeProperty({ collectionId, entityId, propertyId }) {
             where: { tgtEntityId: entityId },
         });
     }
+    return { updated: [property.entityId, property.tgtEntityId] };
 }
 
 export async function associate({ collectionId, entityId, property, tgtEntityId, typeDefinition }) {
@@ -126,27 +130,39 @@ export async function associate({ collectionId, entityId, property, tgtEntityId,
             defaults: property,
         });
     }
+
     // await models.property.bulkCreate(properties);
 }
 
-export async function removeEntity({ id }) {
+export async function removeEntity({ entityId, collectionId }) {
+    let entity = await getEntity({ id: entityId, collectionId });
+    if (!entity) {
+        throw new Error(`You don't have permission to access that entity`);
+    }
+    let targetIds = await models.property.findAll({
+        where: { tgtEntityId: entityId },
+        attributes: ["entityId"],
+        raw: true,
+    });
+    targetIds = targetIds.map((t) => t.entityId);
     await sequelize.transaction(async (t) => {
         await models.entity.destroy({
-            where: { id },
+            where: { id: entityId },
             include: [{ model: models.property }],
             transaction: t,
         });
         // remove properties where this entity is the target
         await models.property.destroy({
-            where: { tgtEntityId: id },
+            where: { tgtEntityId: entityId },
             transaction: t,
         });
         // remove properties associated to this entity
         await models.property.destroy({
-            where: { entityId: id },
+            where: { entityId },
             transaction: t,
         });
     });
+    return { removed: entity, updated: targetIds };
 }
 
 export async function findEntity({ eid, etype, name, hierarchy, collectionId, fuzzy = true }) {
@@ -336,17 +352,17 @@ export async function insertFilesAndFolders({ collectionId, files }) {
         let parent;
         if (!file.parent) {
             //  if parent is undefined attach it to the root dataset
-            parent = (await findEntity({ collectionId, eid: "./", etype: "Dataset" })).pop();
+            parent = await findEntity({ collectionId, eid: "./", etype: "Dataset", fuzzy: false });
+            parent = parent.filter((p) => p.eid === "./" && p.etype === "Dataset").pop();
         } else {
             //  otherwise find the parent and associate to that
-            parent = (
-                await findEntity({
-                    collectionId,
-                    eid: file.parent,
-                    etype: "Dataset",
-                    fuzzy: false,
-                })
-            ).pop();
+            parent = await findEntity({
+                collectionId,
+                eid: file.parent,
+                etype: "Dataset",
+                fuzzy: false,
+            });
+            parent = parent.filter((p) => p.eid === file.parent && p.etype === "Dataset").pop();
         }
         let association = {
             collectionId,
